pyproject.toml
[tool.poetry]
name = "minerva"
version = "2.0.0"
description = "GPT-powered bot built to help students learn software development"
authors = ["Yurij Mikhalevich <yurij@mikhalevi.ch>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12.5"
openai = "^1.43.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.7.0"
python-telegram-bot = "^21.4"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.1"
pytest = "^8.3.2"
flake8 = "^7.1.1"

[tool.poetry.scripts]
minerva = "minerva.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
ignore = ["E125", "E127"]
in-place = true
recursive = true
